pyproject.toml
[project]
name = "labelem"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Patrice Ferlet", email = "metal3d@gmail.com"},
]
dependencies = ["requests>=2.32.5", "pillow>=12.0.0", "torch>=2.9.1", "transformers>=4.57.3", "einop>=0.0.1", "timm>=1.0.22", "streamlit>=1.28.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
